+++ packages/authentication-local/test/fixture.js
@@ -12,7 +12,7 @@ module.exports = (app = feathers()) => {
     entity: 'user',
     service: 'users',
     secret: 'supersecret',
-    jwtStrategies: [ 'local', 'jwt' ],
+    authStrategies: [ 'local', 'jwt' ],
     local: {
       usernameField: 'email',
       passwordField: 'password'