+++ packages/authentication/test/service.test.ts
@@ -27,7 +27,7 @@ describe('authentication/service', () => {
       entity: 'user',
       service: 'users',
       secret: 'supersecret',
-      jwtStrategies: [ 'first' ]
+      authStrategies: [ 'first' ]
     }));
     app.use('/users', memory());
 
@@ -149,7 +149,7 @@ describe('authentication/service', () => {
       const service = app.service('authentication');
       const configuration = service.configuration;
 
-      delete configuration.jwtStrategies;
+      delete configuration.authStrategies;
 
       app.set('authentication', configuration);
 
@@ -161,7 +161,7 @@ describe('authentication/service', () => {
         assert.fail('Should never get here');
       } catch (error) {
         assert.strictEqual(error.name, 'NotAuthenticated');
-        assert.strictEqual(error.message, 'No authentication strategies allowed for creating a JWT (`jwtStrategies`)');
+        assert.strictEqual(error.message, 'No authentication strategies allowed for creating a JWT (`authStrategies`)');
       }
     });
   });
@@ -228,6 +228,22 @@ describe('authentication/service', () => {
       }
     });
 
+    it('throws an error if service name is not set', () => {
+      const otherApp = feathers();
+
+      otherApp.use('/authentication', new AuthenticationService(otherApp, 'authentication', {
+        secret: 'supersecret',
+        authStrategies: [ 'first' ]
+      }));
+
+      try {
+        otherApp.setup();
+        assert.fail('Should never get here');
+      } catch (error) {
+        assert.strictEqual(error.message, `The 'service' option is not set in the authentication configuration`);
+      }
+    });
+
     it('throws an error if entity service does not exist', () => {
       const otherApp = feathers();
 
@@ -235,7 +251,7 @@ describe('authentication/service', () => {
         entity: 'user',
         service: 'users',
         secret: 'supersecret',
-        jwtStrategies: [ 'first' ]
+        authStrategies: [ 'first' ]
       }));
 
       try {