+++ packages/express/test/index.test.js
@@ -19,8 +19,8 @@ describe('@feathersjs/express', () => {
     assert.strictEqual(expressify.default, expressify);
     assert.strictEqual(expressify.original, express);
     assert.strictEqual(typeof expressify.rest, 'function');
-    assert.strictEqual(expressify.errorHandler, require('@feathersjs/errors/handler'));
-    assert.strictEqual(expressify.notFound, require('@feathersjs/errors/not-found'));
+    assert.ok(expressify.notFound);
+    assert.ok(expressify.errorHandler);
   });
 
   it('returns an Express application', () => {
@@ -216,8 +216,8 @@ describe('@feathersjs/express', () => {
       .use('/secureTodos', todoService);
 
     const httpsServer = https.createServer({
-      key: fs.readFileSync(path.join(__dirname, 'resources', 'privatekey.pem')),
-      cert: fs.readFileSync(path.join(__dirname, 'resources', 'certificate.pem')),
+      key: fs.readFileSync(path.join(__dirname, '..', '..', 'tests', 'resources', 'privatekey.pem')),
+      cert: fs.readFileSync(path.join(__dirname, '..', '..', 'tests', 'resources', 'certificate.pem')),
       rejectUnauthorized: false,
       requestCert: false
     }, app).listen(7889);