+++ packages/socketio/test/events.js
@@ -1,6 +1,6 @@
 const assert = require('assert');
 const io = require('socket.io-client');
-const { verify } = require('feathers-commons/lib/test/fixture');
+const { verify } = require('@feathersjs/tests/lib/fixture');
 
 module.exports = function (name, options) {
   const call = (method, ...args) => {